backend/requirements.txt
fastapi
uvicorn
pydantic
google-genai
sentence-transformers
vaderSentiment
numpy
scikit-learn
python-dotenv
spacy